FooterView color logic cleanup
Bug reports showed that when night mode eneded due to schedule, the text color was updated but not the background color.  Just from analyzing the code, this appears to be due to onConfigurationChanged not being called on the footer view, but NSSL.onUiModeChanged was calling into FooterView.setTextColor.  So logically, this seems like a likely fix, even though I can't repro the bug.

Fixes: 197449800
Test: manual theme change continues to update footer colors.
Test: manual language change updates footer text.
Change-Id: If6fdb527eee244561dcd3975d859de93609662a5
(cherry picked from commit 930aebb504b0981152cbcb6c51863b95ddfcea37)