Fix fail verification without cleaning child session data
When a multipackage session is verified to fail, destroyInternal() is
only called for the parent session. However, the child session
doesn't call destroyInternal(). It makes the child session data
occupy the storage space. The device user can't get the storage
space back except to execute factory reset or root the device.

This patch calls destroyInternal() for all of child sessions if the
verification failed.

This patch also creates a new test that needs a rooted device. The
test verifies that the child session stage dir is cleaned.
* using "dumpsys package" to get stage dir
* using "su root" to check the stage dir exist or not so it needs a
  rooted device

Test: TID="PackageManagerServiceHostTests"; \
      TC="com.android.server.pm.test.PackageInstallerSessionTest"; \
      atest "${TID}:${TC}"
Fix: 196522826
Change-Id: Ice257df8dd541c7ca732279ddf0353e6fde83e64