Copy AutofillOptions into derived Contexts
a4158fb fixes missing ContentCaptureOptions in the derived Context.
The same issue should also exist for AutofillOptions in the derived
Contexts. This change copy AutofillOptions into derived Contexts. The
existing AttachedContextActivityTest verifies the derived Contexts.
But this don't really reflact the issue, we will add a test later to
reflect the issue.

Bug: 194542019
Test: atest AttachedContextActivityTest

Change-Id: I45bc87a6de37700de61d4a906001d6a4bec339ca