Register for doze brightness as soon as we can
On devices with under display prox, binned brightness
no longer uses the prox sensor, so let's remove the
brightnessUsesProx config that was only used for
devices with under display prox (to avoid the blinking
white dot).

We now also no longer need to wait on the screen state
to change to dozing states to register for the brightness
sensor since the brightness sensor doesn't check prox.

Test: manual, atest DozeScreenBrightnessTest
Fixes: 194151347
Change-Id: Ie5110585b683accccc1066a06b069072947ffb3f