Stop RBC affecting brightness so drastically
This change resets the brightness short term model when RBC is turned
off or the strength is changed. We add a singular interaction when RBC is turned on, and add
interactions in RBC as normal.

Since users turning RBC in a bright environment could want any
brightness between the given one and MAX, but turning RBC off will often result in
a MAX interaction being logged, if the slider is already at MAX (given
RBC) which is not necessarily what we want.

In order to prevent unwanted changes to the brightness curve, we ignore
interactions that turn RBC off. Turning RBC on seems like a valid
indicator that the user wants the brightness darker, and at a specific
brightness, however turning RBC off is less of an indicator. Changing
the strength of RBC is also a good time to reset the short term model,
however after resetting, we still add an interaction after this.

Bug: 199187829
Test: manual
Change-Id: I6512d9642f2da634aa60af6b5573c792c0a116d5